Changed the type of the parameters

+++ app/src/main/java/team/project/weather/service/WeatherService.java
@@ -3,5 +3,5 @@ package team.project.weather.service;
 import team.project.weather.model.Day;
 
 public interface WeatherService {
-    Day getCurrentDay(float lat,float lon) throws Exception;
+    Day getCurrentDay(double lat,double lon) throws Exception;
 }
